+++ include/Module/Channel/AWGN/Channel_AWGN_LLR.hpp
@@ -38,7 +38,7 @@ public:
 	void set_seed(const int seed);
 
 protected:
-	void _add_noise(const R *X_N, R *Y_N, const int frame_id);
+	void _add_noise(const R *X_N, R *Y_N, const size_t frame_id);
 
 	virtual void deep_copy(const Channel_AWGN_LLR<R>& m);
