+++ include/Module/Channel/User/Channel_user.hpp
@@ -38,9 +38,9 @@ public:
 	static void read_as_binary (const std::string &filename, const int N, std::vector<std::vector<R>>& noise_buffer);
 
 protected:
-	virtual void _add_noise(const R *X_N, R *Y_N, const int frame_id);
+	virtual void _add_noise(const R *X_N, R *Y_N, const size_t frame_id);
 
-	virtual void set_noise(const int frame_id);
+	virtual void set_noise(const size_t frame_id);
 
 private:
 	std::vector<std::vector<R>> noise_buff;
