+++ include/Module/Channel/User/Channel_user_add.hpp
@@ -30,7 +30,7 @@ public:
 	virtual Channel_user_add<R>* clone() const;
 
 protected:
-	virtual void _add_noise(const R *X_N, R *Y_N, const int frame_id);
+	virtual void _add_noise(const R *X_N, R *Y_N, const size_t frame_id);
 };
 }
 }
