+++ include/Module/Decoder/Decoder_HIHO.hpp
@@ -68,9 +68,9 @@ public:
 	int decode_hiho_cw(const B *Y_N, B *V_N, const int frame_id = -1, const bool managed_memory = true);
 
 protected:
-	virtual int _decode_hiho(const B *Y_N, B *V_K, const int frame_id);
+	virtual int _decode_hiho(const B *Y_N, B *V_K, const size_t frame_id);
 
-	virtual int _decode_hiho_cw(const B *Y_N, B *V_N, const int frame_id);
+	virtual int _decode_hiho_cw(const B *Y_N, B *V_N, const size_t frame_id);
 };
 }
 }
