+++ include/Module/Decoder/Polar/SCF/Decoder_polar_SCF_naive_sys.hpp
@@ -30,7 +30,7 @@ public:
 
 protected:
 	void _store   (B *V, bool coded = false) const;
-	bool check_crc(const int frame_id      );
+	bool check_crc(const size_t frame_id   );
 };
 }
 }
